fix: incompatible type

+++ database/migrations/2025_05_20_062528_create_achievement_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('achievement', function (Blueprint $table) {
-            $table->integer("id")->primary();
+            $table->unsignedBigInteger("id")->primary();
             $table->timestamp("upload_at");
             $table->string("competition_type");
             $table->string("competition_name");
